# WIP: use localStorage to cache compiled methods