--- a/main.js
+++ b/main.js
@@ -222,7 +222,7 @@ window.onload = function() {
    }
  };
  document.getElementById("clearCompiledMethodCache").onclick = function() {
-   CompiledMethodCache.clear().then(function() { console.log("cleared compiled method cache") });
+   CompiledMethodCache.clear();
  };
  document.getElementById("printAllExceptions").onclick = function() {
    VM.DEBUG_PRINT_ALL_EXCEPTIONS = !VM.DEBUG_PRINT_ALL_EXCEPTIONS;