--- a/midp/midp.js
+++ b/midp/midp.js
@@ -954,12 +954,9 @@ Native["com/sun/cldc/isolate/Isolate.stop.(II)V"] = function(code, reason) {
         fs.remove("/midlet.jar");
         fs.create("/midlet.jar", new Blob([ data.jarData ]));
 
-        Promise.all([
-            new Promise(function(resolve, reject) {
-                fs.syncStore(resolve);
-            }),
-            CompiledMethodCache.clear(),
-        ]).then(function() {
+        CompiledMethodCache.clear();
+
+        fs.syncStore(function() {
             MIDP.pendingMIDletUpdate = null;
             DumbPipe.close(DumbPipe.open("alert", "Update completed!"));
             DumbPipe.close(DumbPipe.open("reload", {}));