This branch makes the compiled method cache use localstorage.

@mbx I timed a midlet with this branch (using my time-startup branch), and I saw no significant difference in startup performance before vs. after, so I don't think it's worth merging this branch.

```
Before:

"pre-main prep time: 1 ms" relooper.js:4
"time loadApp 1666.2119790000002" main.js:471
"time restore compiled method cache 3794.3570809999997" compiled-method-cache.js:47
"time initFS 3833.5113" fs-init.js:56
"time jvm.startIsolate0 4335.857811000001" main.js:176
"time start 3968.145207" main.js:171
"time start FG MIDlet: 8972.647652" background.js:71
"time DisplayDevice.gainedForeground0 23680.536085" gfx.js:86

"pre-main prep time: 1 ms" relooper.js:4
"time loadApp 1602.907603" main.js:471
"time restore compiled method cache 3184.413227" compiled-method-cache.js:47
"time initFS 3727.479842" fs-init.js:56
"time start 3899.779165" main.js:171
"time jvm.startIsolate0 4154.234425000001" main.js:176
"time start FG MIDlet: 8781.630205000001" background.js:71
"time DisplayDevice.gainedForeground0 21873.225773000002" gfx.js:86

"pre-main prep time: 1 ms" relooper.js:4
"time loadApp 1534.967083" main.js:471
"time restore compiled method cache 1924.2345300000002" compiled-method-cache.js:47
"time initFS 3607.316874" fs-init.js:56
"time start 3756.41078" main.js:171
"time jvm.startIsolate0 4022.5746339999996" main.js:176
"time start FG MIDlet: 8012.515673999999" background.js:71
"time DisplayDevice.gainedForeground0 22770.877075" gfx.js:86

"pre-main prep time: 1 ms" relooper.js:4
"time loadApp 1571.317656" main.js:471
"time restore compiled method cache 3213.298124" compiled-method-cache.js:47
"time initFS 3674.558436" fs-init.js:56
"time start 3909.095102" main.js:171
"time jvm.startIsolate0 4188.787238" main.js:176
"time start FG MIDlet: 8134.717393000001" background.js:71
"time DisplayDevice.gainedForeground0 21957.540512" gfx.js:86

"pre-main prep time: 1 ms" relooper.js:4
"time loadApp 1604.648802" main.js:471
"time restore compiled method cache 3550.899114" compiled-method-cache.js:47
"time initFS 3651.642551" fs-init.js:56
"time start 3781.118905" main.js:171
"time jvm.startIsolate0 4050.7092699999994" main.js:176
"time start FG MIDlet: 8556.261508" background.js:71
"time DisplayDevice.gainedForeground0 21036.94343" gfx.js:86

After:

"pre-main prep time: 1 ms" relooper.js:4
"time loadApp 1765.69526" main.js:471
"time initFS 3796.314061" fs-init.js:56
"time start 3963.27578" main.js:171
"time jvm.startIsolate0 4249.013592" main.js:176
"time start FG MIDlet: 8529.672444999998" background.js:71
"time DisplayDevice.gainedForeground0 23714.964679" gfx.js:86

"time start FG MIDlet: 8690.575778" background.js:71
"pre-main prep time: 1 ms" relooper.js:4
"time loadApp 1887.983436" main.js:471
"time initFS 2755.7358839999997" fs-init.js:56
"time start 2929.260051" main.js:171
"time jvm.startIsolate0 3193.039269" main.js:176
"time start FG MIDlet: 8690.575778" background.js:71
"time DisplayDevice.gainedForeground0 23238.265512" gfx.js:86

"time start 3915.703697" main.js:171
"time jvm.startIsolate0 4189.828748" main.js:176
"pre-main prep time: 1 ms" relooper.js:4
"time loadApp 1632.3781239999998" main.js:471
"time initFS 3761.817029" fs-init.js:56
"time start 3915.703697" main.js:171
"time jvm.startIsolate0 4189.828748" main.js:176
"time start FG MIDlet: 8485.815882" background.js:71
"time DisplayDevice.gainedForeground0 21760.815564" gfx.js:86

"time start FG MIDlet: 8564.767548000002" background.js:71
"pre-main prep time: 1 ms" relooper.js:4
"time loadApp 1774.121926" main.js:471
"time initFS 2750.1417699999997" fs-init.js:56
"time start 2917.701925" main.js:171
"time jvm.startIsolate0 3186.54255" main.js:176
"time start FG MIDlet: 8564.767548000002" background.js:71
"time DisplayDevice.gainedForeground0 22770.248949999997" gfx.js:86

"pre-main prep time: 1 ms" relooper.js:4
"time loadApp 1738.441614" main.js:471
"time initFS 4085.244791" fs-init.js:56
"time start 4237.284737999999" main.js:171
"time jvm.startIsolate0 4530.468176" main.js:176
"time start FG MIDlet: 8842.07583" background.js:71
"time DisplayDevice.gainedForeground0 21026.216972000002" gfx.js:86
```