@andreasgal 2015-02-04
r-. localStorage is evil. Its sync and blocks JS execution and rendering and its particularly bad with e10s. This is architecturally wrong. There has to be a better solution.