@mykmelez 2015-02-04
@andreasgal: We already have a solution for it that uses IndexedDB. But @gregorwagner suggested we try *localStorage* to see if it's any faster, so I created this branch to test that hypothesis.

It doesn't seem to be any faster, however, and there are good reasons to think it'll be slower in some cases (and have the other problems that you note). So I agree that we shouldn't take this.

(I'm looking into speeding up loading of the IndexedDB in another branch that I'll push shortly.)