@andreasgal 2015-02-04
Sounds good. You just gave me a heart attack when I saw localStorage. localStorage and I are not on talking terms after I spent weeks to untangle our dependency on it in Gaia when we were trying to speed up load times there.