@mykmelez 2015-02-04
> Sounds good. You just gave me a heart attack when I saw localStorage. localStorage and I are not on talking terms after I spent weeks to untangle our dependency on it in Gaia when we were trying to speed up load times there.

Note that we do use it to store a single short string:

https://github.com/mozilla/j2me.js/blob/36a3b0af4a4409d7773063e9d55195a6d4ef3755/libs/compiled-method-cache.js#L107

IndexedDB would be overkill for that value. But should we use cookies instead to avoid using localStorage entirely?